refactor(document-preview): split doc ooxml open-document pages and pdf extractors
