+++ src/preview/document/mod.rs
@@ -1,13 +1,16 @@
 mod common;
+mod formats;
 mod metadata;
 
 use self::{
     common::{
-        DOCUMENT_XML_ENTRY_LIMIT_BYTES, extract_zip_document_metadata, first_present_string,
-        format_unix_utc, local_name, parse_xml_text_fields, present_count, present_str,
-        present_string, push_count_stat, push_metadata_field, read_zip_entry,
+        extract_zip_document_metadata, local_name, present_str, read_zip_entry,
         read_zip_entry_bytes_limited, read_zip_entry_limited, resolve_zip_entry_path,
         strip_fragment_identifier, xml_attribute_value,
+    },
+    formats::{
+        extract_doc_metadata, extract_ooxml_metadata, extract_open_document_metadata,
+        extract_pages_metadata, extract_pdf_metadata,
     },
     metadata::{DocumentMetadata, render_document_preview, render_document_preview_lines},
 };
@@ -22,7 +25,6 @@ use std::{
     hash::{Hash, Hasher},
     io::{Read, Write},
     path::{Path, PathBuf},
-    process::Command,
     sync::{Arc, Mutex, OnceLock},
     time::SystemTime,
 };
@@ -35,24 +37,6 @@ const EPUB_COVER_ENTRY_LIMIT_BYTES: usize = 4 * 1024 * 1024;
 const EPUB_SECTION_IMAGE_ENTRY_LIMIT_BYTES: usize = 32 * 1024 * 1024;
 const EPUB_PACKAGE_CACHE_LIMIT: usize = 16;
 const EPUB_ASSET_CACHE_VERSION: usize = 2;
-const DOC_SUMMARY_INFORMATION_STREAM: &str = "/\u{5}SummaryInformation";
-const DOC_PROPERTY_TITLE: u32 = 2;
-const DOC_PROPERTY_SUBJECT: u32 = 3;
-const DOC_PROPERTY_AUTHOR: u32 = 4;
-const DOC_PROPERTY_LAST_SAVED_BY: u32 = 8;
-const DOC_PROPERTY_CREATED: u32 = 12;
-const DOC_PROPERTY_MODIFIED: u32 = 13;
-const DOC_PROPERTY_PAGE_COUNT: u32 = 14;
-const DOC_PROPERTY_WORD_COUNT: u32 = 15;
-const DOC_PROPERTY_CHAR_COUNT: u32 = 16;
-const DOC_PROPERTY_APPLICATION: u32 = 18;
-const VT_I4: u16 = 0x0003;
-const VT_LPSTR: u16 = 0x001E;
-const VT_LPWSTR: u16 = 0x001F;
-const VT_FILETIME: u16 = 0x0040;
-const VT_UI4: u16 = 0x0013;
-const WINDOWS_TICKS_PER_SECOND: u64 = 10_000_000;
-const WINDOWS_TO_UNIX_EPOCH_SECONDS: u64 = 11_644_473_600;
 
 struct EpubPreviewData {
     metadata: DocumentMetadata,
@@ -149,12 +133,6 @@ struct EpubPackageCacheKey {
     modified: Option<(u64, u32)>,
 }
 
-enum DocPropertyValue {
-    Count(u64),
-    Text(String),
-    Timestamp(String),
-}
-
 pub(super) fn build_document_preview(
     path: &Path,
     format: DocumentFormat,
@@ -182,54 +160,6 @@ pub(super) fn build_document_preview(
     }?;
 
     Some(render_document_preview(format, metadata))
-}
-
-fn extract_doc_metadata(path: &Path) -> Option<DocumentMetadata> {
-    File::open(path).ok()?;
-
-    let mut metadata = DocumentMetadata {
-        variant: Some("Legacy binary document".to_string()),
-        ..DocumentMetadata::default()
-    };
-    let mut compound = match cfb::open(path) {
-        Ok(compound) => compound,
-        Err(_) => return Some(metadata),
-    };
-    let stream = match compound.open_stream(DOC_SUMMARY_INFORMATION_STREAM) {
-        Ok(stream) => stream,
-        Err(_) => return Some(metadata),
-    };
-    let mut bytes = Vec::with_capacity(DOCUMENT_XML_ENTRY_LIMIT_BYTES);
-    stream
-        .take(DOCUMENT_XML_ENTRY_LIMIT_BYTES as u64)
-        .read_to_end(&mut bytes)
-        .ok()?;
-    let properties = parse_doc_property_set(&bytes);
-
-    metadata.title = doc_property_text(&properties, DOC_PROPERTY_TITLE);
-    metadata.subject = doc_property_text(&properties, DOC_PROPERTY_SUBJECT);
-    metadata.author = doc_property_text(&properties, DOC_PROPERTY_AUTHOR);
-    metadata.modified_by = doc_property_text(&properties, DOC_PROPERTY_LAST_SAVED_BY);
-    metadata.application = doc_property_text(&properties, DOC_PROPERTY_APPLICATION);
-    metadata.created = doc_property_time(&properties, DOC_PROPERTY_CREATED);
-    metadata.modified = doc_property_time(&properties, DOC_PROPERTY_MODIFIED);
-    push_count_stat(
-        &mut metadata,
-        "Pages",
-        doc_property_count(&properties, DOC_PROPERTY_PAGE_COUNT),
-    );
-    push_count_stat(
-        &mut metadata,
-        "Words",
-        doc_property_count(&properties, DOC_PROPERTY_WORD_COUNT),
-    );
-    push_count_stat(
-        &mut metadata,
-        "Characters",
-        doc_property_count(&properties, DOC_PROPERTY_CHAR_COUNT),
-    );
-
-    Some(metadata)
 }
 
 fn build_epub_preview(path: &Path, section_index: usize) -> Option<PreviewContent> {
@@ -304,191 +234,6 @@ fn render_epub_preview(preview: EpubPreviewData) -> PreviewContent {
         content = content.with_truncation(note);
     }
     content
-}
-
-fn extract_ooxml_metadata<R: Read + std::io::Seek>(
-    archive: &mut ZipArchive<R>,
-    format: DocumentFormat,
-) -> DocumentMetadata {
-    let mut metadata = DocumentMetadata::default();
-
-    let core = read_zip_entry(archive, "docProps/core.xml")
-        .map(|xml| parse_xml_text_fields(&xml))
-        .unwrap_or_default();
-    let app = read_zip_entry(archive, "docProps/app.xml")
-        .map(|xml| parse_xml_text_fields(&xml))
-        .unwrap_or_default();
-
-    metadata.title = present_string(core.get("title"), "Title");
-    metadata.subject = present_string(core.get("subject"), "Subject");
-    metadata.author = present_string(core.get("creator"), "Author");
-    metadata.modified_by = present_string(core.get("lastModifiedBy"), "Modified By");
-    metadata.created = present_string(core.get("created"), "Created");
-    metadata.modified = present_string(core.get("modified"), "Modified");
-    metadata.application = present_string(app.get("Application"), "Application");
-    if let Some(company) = present_string(app.get("Company"), "Company") {
-        metadata.metadata.push(("Company".to_string(), company));
-    }
-
-    match format {
-        DocumentFormat::Docx | DocumentFormat::Docm => {
-            push_count_stat(&mut metadata, "Pages", present_count(app.get("Pages")));
-            push_count_stat(&mut metadata, "Words", present_count(app.get("Words")));
-            push_count_stat(
-                &mut metadata,
-                "Characters",
-                present_count(app.get("Characters")),
-            );
-        }
-        DocumentFormat::Pptx | DocumentFormat::Pptm => {
-            push_count_stat(&mut metadata, "Slides", present_count(app.get("Slides")));
-            push_count_stat(&mut metadata, "Notes", present_count(app.get("Notes")));
-            push_count_stat(
-                &mut metadata,
-                "Hidden Slides",
-                present_count(app.get("HiddenSlides")),
-            );
-            push_count_stat(
-                &mut metadata,
-                "Media Clips",
-                present_count(app.get("MMClips")),
-            );
-        }
-        DocumentFormat::Xlsx | DocumentFormat::Xlsm => {
-            if let Some(manager) = present_string(app.get("Manager"), "Manager") {
-                metadata.metadata.push(("Manager".to_string(), manager));
-            }
-        }
-        _ => {}
-    }
-
-    metadata
-}
-
-fn extract_open_document_metadata<R: Read + std::io::Seek>(
-    archive: &mut ZipArchive<R>,
-    format: DocumentFormat,
-) -> DocumentMetadata {
-    let mut metadata = DocumentMetadata::default();
-    let Some(xml) = read_zip_entry(archive, "meta.xml") else {
-        return metadata;
-    };
-
-    let fields = parse_xml_text_fields(&xml);
-    metadata.title = present_string(fields.get("title"), "Title");
-    metadata.subject = present_string(fields.get("subject"), "Subject");
-    metadata.author = present_string(
-        fields.get("initial-creator").or(fields.get("creator")),
-        "Author",
-    );
-    metadata.created = present_string(fields.get("creation-date"), "Created");
-    metadata.modified = present_string(fields.get("date"), "Modified");
-    metadata.application = present_string(fields.get("generator"), "Application");
-
-    match format {
-        DocumentFormat::Odt => {
-            push_count_stat(
-                &mut metadata,
-                "Pages",
-                present_count(fields.get("page-count")),
-            );
-            push_count_stat(
-                &mut metadata,
-                "Words",
-                present_count(fields.get("word-count")),
-            );
-            push_count_stat(
-                &mut metadata,
-                "Characters",
-                present_count(fields.get("character-count")),
-            );
-        }
-        DocumentFormat::Ods => {
-            push_count_stat(
-                &mut metadata,
-                "Tables",
-                present_count(fields.get("table-count")),
-            );
-            push_count_stat(
-                &mut metadata,
-                "Cells",
-                present_count(fields.get("cell-count")),
-            );
-            push_count_stat(
-                &mut metadata,
-                "Objects",
-                present_count(fields.get("object-count")),
-            );
-        }
-        DocumentFormat::Odp => {
-            push_count_stat(
-                &mut metadata,
-                "Slides",
-                present_count(fields.get("page-count")),
-            );
-            push_count_stat(
-                &mut metadata,
-                "Objects",
-                present_count(fields.get("object-count")),
-            );
-        }
-        _ => {}
-    }
-
-    metadata
-}
-
-fn extract_pages_metadata<R: Read + std::io::Seek>(
-    archive: &mut ZipArchive<R>,
-) -> DocumentMetadata {
-    let mut metadata = DocumentMetadata {
-        application: Some("Apple Pages".to_string()),
-        variant: detect_pages_variant(archive),
-        ..DocumentMetadata::default()
-    };
-
-    let properties = [
-        "Metadata/Properties.plist",
-        "metadata.plist",
-        "QuickLook/Metadata.plist",
-    ]
-    .iter()
-    .find_map(|name| read_zip_entry(archive, name).and_then(|xml| parse_plist_dict(&xml)));
-
-    if let Some(fields) = properties {
-        metadata.title = first_present_string(
-            &fields,
-            &["document-title", "kMDItemTitle", "title", "Title"],
-            "Title",
-        );
-        metadata.subject = first_present_string(
-            &fields,
-            &["subject", "kMDItemDescription", "abstract"],
-            "Subject",
-        );
-        metadata.author = first_present_string(
-            &fields,
-            &["author", "authors", "kMDItemAuthors", "kMDItemAuthor"],
-            "Author",
-        );
-        metadata.created = first_present_string(
-            &fields,
-            &["creationDate", "created", "kMDItemContentCreationDate"],
-            "Created",
-        );
-        metadata.modified = first_present_string(
-            &fields,
-            &[
-                "modificationDate",
-                "modified",
-                "lastOpenedDate",
-                "kMDItemContentModificationDate",
-            ],
-            "Modified",
-        );
-    }
-
-    metadata
 }
 
 fn extract_epub_preview_data<R: Read + std::io::Seek>(
@@ -574,113 +319,6 @@ fn load_epub_package<R: Read + std::io::Seek>(
         cache_epub_package(cache_key, Arc::clone(&cached));
     }
     Some(cached)
-}
-
-fn extract_pdf_metadata(path: &Path) -> Option<DocumentMetadata> {
-    let mut bytes = Vec::with_capacity(256);
-    File::open(path)
-        .ok()?
-        .take(256)
-        .read_to_end(&mut bytes)
-        .ok()?;
-
-    let mut metadata = DocumentMetadata::default();
-    if let Some(version) = parse_pdf_version(&bytes) {
-        metadata.variant = Some(format!("PDF {version}"));
-        metadata
-            .metadata
-            .push(("PDF Version".to_string(), version.to_string()));
-    }
-
-    let output = Command::new("pdfinfo").arg(path).output().ok();
-    let Some(output) = output.filter(|output| output.status.success()) else {
-        return Some(metadata);
-    };
-    let fields = parse_pdfinfo_fields(&String::from_utf8_lossy(&output.stdout));
-    metadata.title = fields
-        .get("Title")
-        .and_then(|value| present_str(value, "Title"));
-    metadata.subject = fields
-        .get("Subject")
-        .and_then(|value| present_str(value, "Subject"));
-    metadata.author = fields
-        .get("Author")
-        .and_then(|value| present_str(value, "Author"));
-    metadata.application = fields
-        .get("Creator")
-        .and_then(|value| present_str(value, "Application"));
-    metadata.created = fields
-        .get("CreationDate")
-        .and_then(|value| present_str(value, "Created"));
-    metadata.modified = fields
-        .get("ModDate")
-        .and_then(|value| present_str(value, "Modified"));
-
-    push_count_stat(
-        &mut metadata,
-        "Pages",
-        fields
-            .get("Pages")
-            .and_then(|value| value.trim().parse().ok()),
-    );
-    push_metadata_field(
-        &mut metadata,
-        "Producer",
-        fields
-            .get("Producer")
-            .and_then(|value| present_str(value, "Producer")),
-    );
-    push_metadata_field(
-        &mut metadata,
-        "Page Size",
-        fields
-            .get("Page size")
-            .and_then(|value| present_str(value, "Page size")),
-    );
-    push_metadata_field(
-        &mut metadata,
-        "Tagged",
-        fields
-            .get("Tagged")
-            .and_then(|value| present_str(value, "Tagged")),
-    );
-    push_metadata_field(
-        &mut metadata,
-        "Encrypted",
-        fields
-            .get("Encrypted")
-            .and_then(|value| present_str(value, "Encrypted")),
-    );
-    push_metadata_field(
-        &mut metadata,
-        "Optimized",
-        fields
-            .get("Optimized")
-            .and_then(|value| present_str(value, "Optimized")),
-    );
-
-    Some(metadata)
-}
-
-fn detect_pages_variant<R: Read + std::io::Seek>(archive: &mut ZipArchive<R>) -> Option<String> {
-    let mut saw_iwa = false;
-    let mut saw_legacy_index = false;
-    for index in 0..archive.len() {
-        let Ok(entry) = archive.by_index(index) else {
-            continue;
-        };
-        let name = entry.name().to_ascii_lowercase();
-        saw_iwa |= name.ends_with(".iwa");
-        saw_legacy_index |= name.ends_with("index.xml") || name.ends_with("index.xml.gz");
-    }
-
-    if saw_iwa {
-        Some("iWork package".to_string())
-    } else if saw_legacy_index {
-        Some("Pages '09 package".to_string())
-    } else {
-        Some("Pages package".to_string())
-    }
 }
 
 fn parse_epub_rootfile_path(xml: &str) -> Option<String> {
@@ -1729,235 +1367,4 @@ fn epub_section_truncation_note() -> String {
         "section excerpt limited to {} KiB",
         EPUB_SECTION_TEXT_LIMIT_CHARS / 1024
     )
-}
-
-fn parse_pdfinfo_fields(output: &str) -> BTreeMap<String, String> {
-    let mut fields = BTreeMap::new();
-    for line in output.lines() {
-        let Some((key, value)) = line.split_once(':') else {
-            continue;
-        };
-        let key = key.trim();
-        let value = value.trim();
-        if key.is_empty() || value.is_empty() {
-            continue;
-        }
-        fields.insert(key.to_string(), value.to_string());
-    }
-    fields
-}
-
-fn parse_pdf_version(bytes: &[u8]) -> Option<&str> {
-    let header = std::str::from_utf8(bytes).ok()?;
-    let header = header.lines().next()?.trim();
-    header.strip_prefix("%PDF-")
-}
-
-fn parse_plist_dict(xml: &str) -> Option<BTreeMap<String, String>> {
-    let mut reader = Reader::from_str(xml);
-    reader.config_mut().trim_text(true);
-    let mut fields = BTreeMap::new();
-    let mut pending_key: Option<String> = None;
-    let mut current_tag: Option<String> = None;
-    let mut current_array_key: Option<String> = None;
-    let mut array_values = Vec::new();
-
-    loop {
-        match reader.read_event() {
-            Ok(Event::Start(event)) => {
-                let tag = local_name(event.name().as_ref());
-                if tag == "array" {
-                    current_array_key = pending_key.take();
-                    array_values.clear();
-                }
-                current_tag = Some(tag);
-            }
-            Ok(Event::Empty(event)) => {
-                let tag = local_name(event.name().as_ref());
-                if tag == "true" || tag == "false" {
-                    let Some(key) = pending_key.take() else {
-                        continue;
-                    };
-                    fields.insert(key, tag);
-                }
-                current_tag = None;
-            }
-            Ok(Event::Text(text)) => {
-                let Ok(value) = text.decode() else {
-                    continue;
-                };
-                let value = value.trim();
-                if value.is_empty() {
-                    continue;
-                }
-                match current_tag.as_deref() {
-                    Some("key") => pending_key = Some(value.to_string()),
-                    Some("string") | Some("date") | Some("integer") | Some("real") => {
-                        if current_array_key.is_some() {
-                            array_values.push(value.to_string());
-                        } else if let Some(key) = pending_key.take() {
-                            fields.insert(key, value.to_string());
-                        }
-                    }
-                    _ => {}
-                }
-            }
-            Ok(Event::End(event)) => {
-                let tag = local_name(event.name().as_ref());
-                if tag == "array"
-                    && let Some(key) = current_array_key.take()
-                    && !array_values.is_empty()
-                {
-                    fields.insert(key, array_values.join(", "));
-                    array_values.clear();
-                }
-                current_tag = None;
-            }
-            Ok(Event::Eof) | Err(_) => break,
-            _ => {}
-        }
-    }
-
-    if fields.is_empty() {
-        None
-    } else {
-        Some(fields)
-    }
-}
-
-fn parse_doc_property_set(bytes: &[u8]) -> BTreeMap<u32, DocPropertyValue> {
-    let mut properties = BTreeMap::new();
-    let Some(section_count) = read_u32(bytes, 28) else {
-        return properties;
-    };
-    if section_count == 0 {
-        return properties;
-    }
-    let Some(section_offset) = read_u32(bytes, 44).map(|offset| offset as usize) else {
-        return properties;
-    };
-    if section_offset >= bytes.len() {
-        return properties;
-    }
-
-    let section = &bytes[section_offset..];
-    let Some(property_count) = read_u32(section, 4).map(|count| count as usize) else {
-        return properties;
-    };
-    for index in 0..property_count {
-        let entry_offset = 8 + index * 8;
-        let Some(property_id) = read_u32(section, entry_offset) else {
-            continue;
-        };
-        let Some(value_offset) = read_u32(section, entry_offset + 4).map(|offset| offset as usize)
-        else {
-            continue;
-        };
-        if let Some(value) = parse_doc_property_value(section, value_offset) {
-            properties.insert(property_id, value);
-        }
-    }
-
-    properties
-}
-
-fn parse_doc_property_value(section: &[u8], offset: usize) -> Option<DocPropertyValue> {
-    let value_type = read_u16(section, offset)?;
-    match value_type {
-        VT_I4 | VT_UI4 => {
-            read_u32(section, offset + 4).map(|value| DocPropertyValue::Count(value as u64))
-        }
-        VT_LPSTR => parse_lpstr(section, offset + 4).map(DocPropertyValue::Text),
-        VT_LPWSTR => parse_lpwstr(section, offset + 4).map(DocPropertyValue::Text),
-        VT_FILETIME => parse_filetime(section, offset + 4).map(DocPropertyValue::Timestamp),
-        _ => None,
-    }
-}
-
-fn parse_lpstr(bytes: &[u8], offset: usize) -> Option<String> {
-    let length = read_u32(bytes, offset)? as usize;
-    if length == 0 {
-        return None;
-    }
-    let slice = bytes.get(offset + 4..offset + 4 + length)?;
-    let content = slice.strip_suffix(&[0]).unwrap_or(slice);
-    let value = String::from_utf8(content.to_vec()).ok()?;
-    let value = value.trim();
-    (!value.is_empty()).then(|| value.to_string())
-}
-
-fn parse_lpwstr(bytes: &[u8], offset: usize) -> Option<String> {
-    let length = read_u32(bytes, offset)? as usize;
-    if length == 0 {
-        return None;
-    }
-    let byte_len = length.checked_mul(2)?;
-    let slice = bytes.get(offset + 4..offset + 4 + byte_len)?;
-    let mut units = Vec::with_capacity(length);
-    for chunk in slice.chunks_exact(2) {
-        units.push(u16::from_le_bytes([chunk[0], chunk[1]]));
-    }
-    if let Some(0) = units.last().copied() {
-        units.pop();
-    }
-    let value = String::from_utf16(&units).ok()?;
-    let value = value.trim();
-    (!value.is_empty()).then(|| value.to_string())
-}
-
-fn parse_filetime(bytes: &[u8], offset: usize) -> Option<String> {
-    let ticks = read_u64(bytes, offset)?;
-    if ticks < WINDOWS_TO_UNIX_EPOCH_SECONDS * WINDOWS_TICKS_PER_SECOND {
-        return None;
-    }
-    let unix_seconds =
-        (ticks / WINDOWS_TICKS_PER_SECOND).checked_sub(WINDOWS_TO_UNIX_EPOCH_SECONDS)?;
-    format_unix_utc(unix_seconds)
-}
-
-fn read_u16(bytes: &[u8], offset: usize) -> Option<u16> {
-    let slice = bytes.get(offset..offset + 2)?;
-    Some(u16::from_le_bytes([slice[0], slice[1]]))
-}
-
-fn read_u32(bytes: &[u8], offset: usize) -> Option<u32> {
-    let slice = bytes.get(offset..offset + 4)?;
-    Some(u32::from_le_bytes([slice[0], slice[1], slice[2], slice[3]]))
-}
-
-fn read_u64(bytes: &[u8], offset: usize) -> Option<u64> {
-    let slice = bytes.get(offset..offset + 8)?;
-    Some(u64::from_le_bytes([
-        slice[0], slice[1], slice[2], slice[3], slice[4], slice[5], slice[6], slice[7],
-    ]))
-}
-
-fn doc_property_text(
-    properties: &BTreeMap<u32, DocPropertyValue>,
-    property_id: u32,
-) -> Option<String> {
-    match properties.get(&property_id) {
-        Some(DocPropertyValue::Text(value)) => present_str(value, ""),
-        _ => None,
-    }
-}
-
-fn doc_property_time(
-    properties: &BTreeMap<u32, DocPropertyValue>,
-    property_id: u32,
-) -> Option<String> {
-    match properties.get(&property_id) {
-        Some(DocPropertyValue::Timestamp(value)) => present_str(value, ""),
-        _ => None,
-    }
-}
-
-fn doc_property_count(
-    properties: &BTreeMap<u32, DocPropertyValue>,
-    property_id: u32,
-) -> Option<u64> {
-    match properties.get(&property_id) {
-        Some(DocPropertyValue::Count(value)) => Some(*value),
-        _ => None,
-    }
 }
